+++ include/nfx/Graphics.h
@@ -92,6 +92,7 @@
 #include "nfx/graphics/math/Mat3.h"
 #include "nfx/graphics/math/Mat4.h"
 #include "nfx/graphics/math/Quat.h"
+#include "nfx/graphics/math/Vec2.h"
 #include "nfx/graphics/math/Vec3.h"
 
 // Tools